Nav, Core and FXUI doc finished (#29 #34 #35 #36)
The nav-bar is now included in every .md and can be used to
navigate the project more easily.

The core-module documentation now includes module
content with descriptions.

The FXUI-module documentation now includes module
content with descriptions.

There might be need to update documentation as the
deliverable is not yet finished.