User class and serializing implemented (#58 #61)
The user-class should now be implemented with json, with
serilizing and dezerializing.

I needed to change a lot of other classes to get rid of errors.
They have all been tagged with a TODO.

I still need to test the implementation, as my
apple silicon is incompatible.
Expect a refactor commit soon.