Implemented info-page for each event (#57)
An informational page that displays the attributes of an event.
The user can register to or deregister from an event on this page.
Some new methods have been added to the IOUtil file in module core,
package json. The event cells now link to this page. Owner of event
functionality has not yet been implemented.